+++ kernel/src/sys/idt.zig
@@ -41,7 +41,7 @@ pub const IDT = struct {
 
     pub fn load(self: *@This()) void {
         logger.info("Set interrupt handlers", .{});
-        comptime var i: usize = 0;
+        comptime var i: u64 = 0;
         inline while (i < 256) : (i += 1) {
             const handler = comptime ivt.makeHandler(i);
             self.entries[i] = IDTEntry.make(@intFromPtr(handler), 0, interrupt_gate);
